+++ CHANGELOG.md
@@ -46,6 +46,10 @@
     error message more descriptive when the Set-TargetResource function
     calls the Test-TargetResource function to verify the desired
     state. *This was done prior to adding full en-US localization.*
+  - Fixed ([issue #1258](https://github.com/PowerShell/SqlServerDsc/issues/1258)).
+    When initializing Reporting Services, there is no need to execute `InitializeReportServer`
+    CIM method, since executing `SetDatabaseConnection` CIM method initializes
+    Reporting Services.
 - Changes to SqlServerLogin
   - Added en-US localization ([issue #615](https://github.com/PowerShell/SqlServerDsc/issues/615)).
   - Added unit tests to improved code coverage.